clean up

+++ bstrkr.mobile/bstrkr.core/Consts/AppConsts.cs
@@ -10,6 +10,6 @@ namespace bstrkr.core.consts
 		public const double MaxDistance = 50.0;
 
 		public static GeoPoint DefaultLocation = new GeoPoint(55.7503798d, 37.6182293d);
-		public const double DefaultZoom = 14d;
+		public const float DefaultZoom = 14.0f;
 	}
 }
